+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ## 1.0.0
 
 Initial release.  Provides basic data source functionality of fetching saved/scheduled search results from Cribl.
+
+## 1.1.0
+
+- Add support for using dashboard variables in Cribl Search queries.
+- Add support for Grafana alerting.